Use the real destination IP for send()
The destination IP passed to send() controls the routing decision.
Using the real destination IP is important if the route for the dummy IP
(1.0.0.0) differs from the route for the tunnel peer.  This is most
likely to happen if the tunnel peer is connected via something other
than the default route, but it could also happen in theory if 1.0.0.0
was in the routing table.